+++ packages/patrol/android/src/main/kotlin/pl/leancode/patrol/Automator.kt
@@ -61,7 +61,7 @@ private fun fromUiObject2(obj: UiObject2): AndroidNativeView {
             minX = bounds.left.toDouble(),
             minY = bounds.top.toDouble(),
             maxX = bounds.right.toDouble(),
-            maxY = bounds.top.toDouble()
+            maxY = bounds.bottom.toDouble()
         ),
         visibleCenter = Point2D(
             x = center.x.toDouble(),
@@ -111,8 +111,6 @@ class Automator private constructor() {
         configurator.waitForIdleTimeout = 5000
         configurator.keyInjectionDelay = 50
 
-        configurator.uiAutomationFlags = UiAutomation.FLAG_DONT_SUPPRESS_ACCESSIBILITY_SERVICES
-
         Logger.i("Timeout: $timeoutMillis ms")
         Logger.i("Android UiAutomator configuration:")
         Logger.i("\twaitForSelectorTimeout: ${configurator.waitForSelectorTimeout} ms")
@@ -121,7 +119,6 @@ class Automator private constructor() {
         Logger.i("\tactionAcknowledgmentTimeout: ${configurator.actionAcknowledgmentTimeout} ms")
         Logger.i("\tscrollAcknowledgmentTimeout: ${configurator.scrollAcknowledgmentTimeout} ms")
         Logger.i("\ttoolType: ${configurator.toolType}")
-        Logger.i("\tuiAutomationFlags: ${configurator.uiAutomationFlags}")
     }
 
     private fun executeShellCommand(cmd: String) {